--- src/users/dto/UsersCreateDTO.ts
@@ -1,7 +1,0 @@
-export class UsersCreateDTO {
-  name: string;
-  lastName: string;
-  email: string;
-  password: string;
-  phone: string;
-}
