+++ src/users/dto/create-user.dto.ts
@@ -1,1 +1,7 @@
-export class CreateUserDto {}
+export class UsersCreateDTO {
+  name: string;
+  lastName: string;
+  email: string;
+  password: string;
+  phone: string;
+}
